adding stronghold card and making cards appear at bottom of the screen to start. also renaming a couple of files.
