+++ cardgame/cards/general_card.py
